Update activations.py
if the input to the sparsemax has dimensions with size=1, they are squeezed by accident, it seems.

e.g.
attn = torch.rand((64, 1, 6))
sparsemax(attn, dim=2)

in the backward pass, supp_size will be (64x1x1), and will be squeezed to (64) instead of (64x1), in line 163. This creates a weirdly shaped v_hat, but doesn't break the code immediately.

In the code I was running, this would only cause the code to break later on after a few training epochs (due to nans in gradients). After applying this change, the training seems stabler and doesn't break